Made the messageRouter configurable
