+++ digiwf-email-integration/src/main/java/io/muenchendigital/digiwf/email/integration/api/streaming/MessageProcessor.java
@@ -1,13 +1,11 @@
-package io.muenchendigital.digiwf.email.integration.infrastructure.streaming;
+package io.muenchendigital.digiwf.email.integration.api.streaming;
 
 import io.muenchendigital.digiwf.email.integration.domain.exception.MissingInformationMailException;
 import io.muenchendigital.digiwf.email.integration.domain.model.Mail;
 import io.muenchendigital.digiwf.email.integration.domain.service.MailingService;
-import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.infrastructure.RoutingCallback;
 import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.service.CorrelateMessageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cloud.function.context.MessageRoutingCallback;
 import org.springframework.context.annotation.Bean;
 import org.springframework.mail.MailException;
 import org.springframework.messaging.Message;
@@ -23,24 +21,14 @@ import java.util.function.Consumer;
 @RequiredArgsConstructor
 public class MessageProcessor {
 
-    public static final String TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS = "sendMailFromEventBus";
+
     private final MailingService mailingService;
     private final CorrelateMessageService correlateMessageService;
     private static final String MAIL_SENT_STATUS = "mailSentStatus";
 
     /**
-     * Override the custom router of the digiwf-spring-cloudstream-utils. We only have one type we need to map.
-     * @return the custom router
-     */
-    @Bean
-    public MessageRoutingCallback mailRouter() {
-        final Map<String, String> typeMappings = new HashMap<>();
-        typeMappings.put(TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS, TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS);
-        return new RoutingCallback(typeMappings);
-    }
-
-    /**
      * All messages from the route "sendMailFromEventBus" go here.
+     *
      * @return the consumer
      */
     @Bean
@@ -61,8 +49,9 @@ public class MessageProcessor {
 
     /**
      * Function to emit a reponse using the correlateMessageService of digiwf-spring-cloudstream-utils
+     *
      * @param messageHeaders The MessageHeaders of the incoming message you want to correlate your answer to
-     * @param status true when the e-mail has been sent, false when an error occured
+     * @param status         true when the e-mail has been sent, false when an error occured
      */
     public void emitResponse(final MessageHeaders messageHeaders, final boolean status) {
         final Map<String, Object> correlatePayload = new HashMap<>();
