+++ digiwf-email-integration/src/main/java/io/muenchendigital/digiwf/email/integration/configuration/MailConfiguration.java
@@ -1,15 +1,20 @@
 package io.muenchendigital.digiwf.email.integration.configuration;
 
+import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.infrastructure.RoutingCallback;
+import org.springframework.cloud.function.context.MessageRoutingCallback;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
 
 import javax.mail.MessagingException;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
 
 @Configuration
 public class MailConfiguration {
+
+    public static final String TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS = "sendMailFromEventBus";
 
     public JavaMailSender getJavaMailSender(
             final String host,
@@ -33,5 +38,15 @@ public class MailConfiguration {
         return mailSender;
     }
 
+    /**
+     * Override the custom router of the digiwf-spring-cloudstream-utils. We only have one type we need to map.
+     *
+     * @return the custom router
+     */
+    public MessageRoutingCallback mailRouter() {
+        final Map<String, String> typeMappings = new HashMap<>();
+        typeMappings.put(TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS, TYPE_HEADER_SEND_MAIL_FROM_EVENT_BUS);
+        return new RoutingCallback(typeMappings);
+    }
 
 }
